+++ src/Contract.sol
@@ -2,5 +2,5 @@
 pragma solidity 0.8.13;
 
 contract Contract {
-    mapping(address => uint) doSomething;
+    mapping(address => uint256) doSomething;
 }
